test: one test_support module for the shared scaffolding
