+++ src/capture/af_packet.rs
@@ -384,9 +384,9 @@ mod tests {
     use std::time::{Duration, Instant};
 
     use super::*;
-    use crate::capture::{Tun, loopback_lock, open_or_skip};
     use crate::net::frame;
     use crate::net::mac::MacAddr;
+    use crate::test_support::{Tun, loopback_lock, open_or_skip};
 
     /// How long a live tun test waits for a packet to cross the device.
     const WAIT_BUDGET: Duration = Duration::from_secs(2);
