+++ src/config/tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 
 fn from_toml(text: &str) -> Result<Config, ConfigError> {
