+++ src/net/http/framing/tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 
 impl RewritePolicy {
