+++ src/reactor/tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 use std::cell::{Cell, RefCell};
 use std::io::Write;
